def_buffer returns a c_style buffer
Adds `__getitem__`, `__setitem__` to accept tuples in the [k,j,i], c_style
order.
The existing accessors with 'Points' keep the [i,j,k], f_style order.

Also add constructors from 'py::arrays' which automatically detect the
order of the input numpy array ('C' or 'F').
The existing constructor from a `py::buffer` doesn't have a default order
parameter, so user needs to be explicit about it.

TODO: Next step, to be consistent with the tuple accessors, is to also
provide the user to specify lower_bounds with tuples in the c-style.
Right now we require a Point, however to reduce confusion, the name of
the argument has been changed from `lower_bound` to `lower_bound_ijk`